Fix some typos, still to fix the attachment so it is cleanly separated from Uploads when we hook uploads module. That way we can eventually have an attachment as opposed to upload field in same form. Just handle it ourselves for now as we need it working, now.

+++ xaruser/contactus.php
@@ -336,7 +336,9 @@ function sitecontact_user_contactus($args)
         if (!xarModAPIFunc('mail','admin','sendhtmlmail', $args))return;
     }
     if (isset($attachpath) && !empty($attachpath)){
-        unlink("{$attachpath}");
+        if (file_exists($attachpath)) {
+            unlink("{$attachpath}");
+        }
     }
     /* Set the theme comments back */
     xarModSetVar('themes','ShowTemplates',$themecomments);
